pyproject.toml
[project]
name = "paper_radar"
version = "0.1.0"
description = "TODO"
requires-python = ">=3.12"
dependencies = [
    "coding-cli-runtime>=0.5.1",
    "torch",
    "torchvision",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sdk_test"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
